+++ src/FbActivityPlugin.php
@@ -33,6 +33,9 @@ class FbActivityPlugin implements Plugin
 
     public static function get(): static
     {
-        return filament(app(static::class)->getId());
+        /** @var static $plugin */
+        $plugin = filament(app(static::class)->getId());
+
+        return $plugin;
     }
 }