+++ tests/Services/Podcast.php
@@ -5,14 +5,14 @@ namespace Mortezamasumi\FbActivity\Tests\Services;
 use Illuminate\Database\Eloquent\Attributes\UseFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Spatie\Activitylog\Traits\LogsActivity;
 use Spatie\Activitylog\LogOptions;
+use Spatie\Activitylog\Traits\LogsActivity;
 
 #[UseFactory(PodcastFactory::class)]
 class Podcast extends Model
 {
-    use LogsActivity;
     use HasFactory;
+    use LogsActivity;
 
     protected $guarded = [];